:hammer: Refactor bottom solid components to support multiplatform usage
