+++ app/src/commonMain/kotlin/com/crosspaste/ui/paste/FileBottomSolid.kt
@@ -13,15 +13,41 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.style.TextOverflow
+import com.crosspaste.i18n.GlobalCopywriter
 import com.crosspaste.ui.theme.AppUIColors
 import com.crosspaste.ui.theme.AppUIFont
 import com.crosspaste.ui.theme.AppUIFont.bottomSolidTitleTextStyle
 import com.crosspaste.ui.theme.AppUISize.tiny4X
+import com.crosspaste.utils.getFileUtils
+import com.crosspaste.utils.safeIsDirectory
+import okio.Path
 
 data class FileDisplayInfo(
     val title: String?,
     val subtitle: String,
 )
+
+fun getFileDisplayInfo(
+    files: List<Path>,
+    copywriter: GlobalCopywriter,
+): FileDisplayInfo? {
+    if (files.isEmpty()) return null
+
+    if (files.size > 1) {
+        val subtitle = files.joinToString(", ") { it.name }
+        return FileDisplayInfo(null, subtitle)
+    }
+
+    val file = files[0]
+    val title = file.name
+
+    if (file.safeIsDirectory) {
+        return FileDisplayInfo(title, copywriter.getText("folder"))
+    }
+
+    val subtitle = getFileUtils().formatBytes(file.toFile().length())
+    return FileDisplayInfo(title, subtitle)
+}
 
 @Composable
 fun FileBottomSolid(
